Fix rubocop errors and create .vim/swap folder

+++ connect.rb
@@ -11,24 +11,20 @@ files = %w(
   .config/i3status/config
   .vim/indent/ruby.vim
 )
-dirs = files.map {|n| File.dirname n}.select {|d| d != '.'}
+dirs = files.map { |n| File.dirname n }.select { |d| d != '.' }
+dirs << '.vim/swap'
 
 def mkdir(name)
-  if not File.exist? "#{Dir.home}/#{name}"
-    FileUtils.mkdir_p(File.join(Dir.home, name))
-  end
+  full_name = "#{Dir.home}/#{name}"
+  FileUtils.mkdir_p(File.join(Dir.home, name)) unless File.exist? full_name
 end
 
 def link_file(name)
   home_name = "#{Dir.home}/#{name}"
-  if File.symlink? home_name
-    File.unlink home_name
-  end
-  if File.exist? home_name
-    File.delete home_name
-  end
+  File.unlink home_name if File.symlink? home_name
+  File.delete home_name if File.exist? home_name
   File.symlink("#{DOT_DIR}/#{name}", home_name)
 end
 
-dirs.each {|name| mkdir name}
-files.each {|name| link_file name}
+dirs.each { |name| mkdir name }
+files.each { |name| link_file name }
